refactor so that top and bottom surfaces are individually specified.
This allows explicitly asking for "blank" surfaces, and ensures that
double-tracked pieces are blank on sides where the double-track connector
would be improperly mirrored.  In the future I might add "fast" surfaces,
which carve out deep 45-degree channels on the unused side to reduce the
volume of the piece and allow faster printing.